feat: Ajustando adicionar e emprestar livros

+++ src/main/java/com/librarymanager/repository/LivroRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.librarymanager.model.Livro;
 
 @Repository
-public interface LivroRepository extends JpaRepository<Livro, Integer> {
-    
+public interface LivroRepository extends JpaRepository<Livro, Long> { // Alterado para Long
 }
